ログインしているユーザーとprototypeを作成したユーザーが同一であるかを判別

+++ app/controllers/prototypes_controller.rb
@@ -42,6 +42,9 @@ class PrototypesController < ApplicationController
 
   def destroy 
     prototype = Prototype.find(params[:id])
+    unless user_signed_in? && current_user.id == prototype.user_id
+      redirect_to root_path
+    end
     prototype.destroy
     redirect_to root_path
   end
